verifying all inputs

+++ actions/action_delete_property.php
@@ -1,8 +1,9 @@
 <?php
 
     include_once('../includes/session.php');
-    include_once('../debug/debug.php');
     include_once('../database/db_property.php');
+    include_once('../includes/input_validation.php');
+    include_once('../includes/redirect.php');
 
 
     $username = $_SESSION['username'];
@@ -12,16 +13,16 @@
         die(header('Location: ../pages/login.php'));
     }
 
-    // TODO: check if the user that is logged in is the owner of the property,
-    // if not set error message and die
 
-    // TODO: check if the id is a number (if it is valid)
     $property_id = $_GET['id'];
-    console_log($property_id);
+    $info = get_property_info($property_id);
+
+    if($info['owner'] != $username){
+        die(redirect('error','You cannot delete other user property'));
+    }
     delete_property($property_id);
 
     $_SESSION['messages'][] = array('type' => 'success', 'content' => 'Property deleted.');
-    //TODO: this must be change ( instead of index we must redirect to the profile page, altough it's not working)
     header("Location: ../pages/manage_properties.php");
 
 ?>
